# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kura Kaani API Documentation",
        "description": "Kura Kaani API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/me": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get Authenticated User Profile",
                "operationId": "48bab3fa17dab7d2aface8905aa88f58",
                "responses": {
                    "200": {
                        "description": "Success response when profile data is fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "properties": {
                                                "head": {
                                                    "type": "string",
                                                    "example": "Data Fetched"
                                                },
                                                "body": {
                                                    "type": "string",
                                                    "example": "Profile fetched successfully"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error response when profile data fetch fails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "properties": {
                                                "head": {
                                                    "type": "string",
                                                    "example": "Failed to Fetch"
                                                },
                                                "body": {
                                                    "type": "string",
                                                    "example": "Profile fetch failed"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/update/{id}": {
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update User Profile",
                "operationId": "878717998c524db474fb041d70bb1521",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Profile data to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response when profile is updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "properties": {
                                                "head": {
                                                    "type": "string",
                                                    "example": "Profile Updated"
                                                },
                                                "body": {
                                                    "type": "string",
                                                    "example": "Profile updated successfully"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error response when profile update fails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "properties": {
                                                "head": {
                                                    "type": "string",
                                                    "example": "Update Failed"
                                                },
                                                "body": {
                                                    "type": "string",
                                                    "example": "Profile update failed"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, Springfield, IL"
                    },
                    "gender": {
                        "type": "string",
                        "example": "male"
                    },
                    "dob": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "ProfileUpdateRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "phone_number",
                    "address",
                    "gender",
                    "dob"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    "dob": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        }
    }
}